% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readFasta}
\alias{readFasta}
\title{Read a fasta file.}
\usage{
readFasta(infile)
}
\arguments{
\item{infile}{`character(1)` with the name of the fasta file to read.}
}
\value{
an instance of \code{DnaSeq} or \code{RnaSeq}.
}
\description{
Reads sequences data in fasta and create \code{DnaSeq} and \code{RnaSeq}
instances.
}
\details{
This funtion reads DNA and RNA fasta files and generates valid
\code{"DnaSeq"} and \code{"RnaSeq"} instances.
}
\examples{
## 2 fasta files
f <- dir(system.file("extdata", package = "sequences"),
         pattern = "fasta", full.names = TRUE)
f

## read first fasta file
x <- readFasta(f[1])
x
}
\seealso{
\code{\linkS4class{GenericSeq}}, \code{\linkS4class{DnaSeq}} and
    \code{\linkS4class{RnaSeq}}.
}
\author{
Laurent Gatto
}
\keyword{IO,}
\keyword{file}
